+++ ui/src/Components/MainModal/Configuration/MultiGridConfiguration.test.tsx
@@ -72,17 +72,18 @@ describe("<MultiGridConfiguration />", () => {
     });
     const tree = ExpandSortLabelSuggestions();
     const options = tree.find("div.react-select__option");
-    expect(options).toHaveLength(4);
+    expect(options).toHaveLength(5);
     expect(options.at(0).text()).toBe("Disable multi-grid");
-    expect(options.at(1).text()).toBe("@alertmanager");
-    expect(options.at(2).text()).toBe("@cluster");
-    expect(options.at(3).text()).toBe("@receiver");
+    expect(options.at(1).text()).toBe("Automatic selection");
+    expect(options.at(2).text()).toBe("@alertmanager");
+    expect(options.at(3).text()).toBe("@cluster");
+    expect(options.at(4).text()).toBe("@receiver");
   });
 
   it("clicking on a label option updates settingsStore", () => {
     const tree = ExpandSortLabelSuggestions();
     const options = tree.find("div.react-select__option");
-    options.at(5).simulate("click");
+    options.at(6).simulate("click");
     expect(settingsStore.multiGridConfig.config.gridLabel).toBe("job");
   });